refactor: replace 1UL with 1ULL for consistent block size calculations across multiple files

+++ include/register_blocked_BF_64bit.h
@@ -18,7 +18,7 @@ public:
 	explicit RegisterBlockedBF64Bit(size_t n_key, uint32_t n_bits_per_key) {
 		num_blocks = ((n_key * n_bits_per_key) >> 6) + 1;
 		num_blocks_log = static_cast<uint32_t>(std::log2(num_blocks)) + 1;
-		num_blocks = std::min(1UL << num_blocks_log, MAX_NUM_BLOCKS);
+		num_blocks = std::min(1ULL << num_blocks_log, MAX_NUM_BLOCKS);
 
 		blocks.resize(num_blocks);
 		std::cout << "BF Size: " << num_blocks * 8 / 1024 << " KiB\n";
@@ -54,8 +54,8 @@ public:
 	}
 
 private:
-	size_t num_blocks;
-	size_t num_blocks_log;
+	uint64_t num_blocks;
+	uint64_t num_blocks_log;
 	std::vector<uint64_t> blocks;
 };
 } // namespace bloom_filters
